added sqlite db
Connected to a sqlite database and made a class that manages the database connection and commands. Made the player database that holds info of the player's position. Use this position from the database when launching main game and when the final position of the player is computed, update the values into the database