added shooting rudimentary
added the initial phase of shooting, where allocating and deallocating bullets was done, they are generated in the middle of the screen. No animation when shooting was done. The bullets count is not on the screen yet, this is the next phase, no collisions. Deallocating happens when the bullets exit the screen.