shooting done
shooting is now pretty good. Divided the sprite into 2 others, the bottom and the top part to handle the different animations. There are 3 sprites for the player, but everything is based on the original one which is not drawn at all on the screen, only updated as a lot of things depend on it and deleting it would have caused a lot of issues. Now, it comes to making a new world and implementing collisions.